nonebot_plugin_tetris_stats/utils/render/schemas/v1/base.py
from pydantic import BaseModel

from ....typedefs import Number
from ..base import HistoryData


class History(BaseModel):
    data: list[HistoryData]
    split_interval: Number
    min_value: Number
    max_value: Number
    offset: Number
